+++ acvm/src/pwg.rs
@@ -7,15 +7,19 @@ use acir::{
 };
 use std::collections::BTreeMap;
 
+use self::arithmetic::ArithmeticSolver;
+
 // arithmetic
 pub mod arithmetic;
 // Directives
 pub mod directives;
-// blackbox functions
+// black box functions
+pub mod block;
 pub mod hash;
 pub mod logic;
 pub mod range;
 pub mod signature;
+pub mod sorting;
 
 // Returns the concrete value for a particular witness
 // If the witness has no assignment, then
@@ -26,40 +30,46 @@ pub fn witness_to_value(
 ) -> Result<&FieldElement, OpcodeResolutionError> {
     match initial_witness.get(&witness) {
         Some(value) => Ok(value),
-        None => Err(OpcodeResolutionError::OpcodeNotSolvable(
-            OpcodeNotSolvable::MissingAssignment(witness.0),
-        )),
+        None => Err(OpcodeNotSolvable::MissingAssignment(witness.0).into()),
     }
 }
+
 // TODO: There is an issue open to decide on whether we need to get values from Expressions
 // TODO versus just getting values from Witness
 pub fn get_value(
     expr: &Expression,
     initial_witness: &BTreeMap<Witness, FieldElement>,
 ) -> Result<FieldElement, OpcodeResolutionError> {
-    let mut result = expr.q_c;
+    let expr = ArithmeticSolver::evaluate(expr, initial_witness);
+    match expr.to_const() {
+        Some(value) => Ok(value),
+        None => {
+            Err(OpcodeResolutionError::OpcodeNotSolvable(OpcodeNotSolvable::MissingAssignment(
+                ArithmeticSolver::any_witness_from_expression(&expr).unwrap().0,
+            )))
+        }
+    }
+}
 
-    for term in &expr.linear_combinations {
-        let coefficient = term.0;
-        let variable = term.1;
+// Inserts `value` into the initial witness map
+// under the key of `witness`.
+// Returns an error, if there was already a value in the map
+// which does not match the value that one is about to insert
+fn insert_value(
+    witness: &Witness,
+    value_to_insert: FieldElement,
+    initial_witness: &mut BTreeMap<Witness, FieldElement>,
+) -> Result<(), OpcodeResolutionError> {
+    let optional_old_value = initial_witness.insert(*witness, value_to_insert);
 
-        // Get the value assigned to that variable
-        let assignment = *witness_to_value(initial_witness, variable)?;
+    let old_value = match optional_old_value {
+        Some(old_value) => old_value,
+        None => return Ok(()),
+    };
 
-        result += coefficient * assignment;
+    if old_value != value_to_insert {
+        return Err(OpcodeResolutionError::UnsatisfiedConstrain);
     }
 
-    for term in &expr.mul_terms {
-        let coefficient = term.0;
-        let lhs_variable = term.1;
-        let rhs_variable = term.2;
-
-        // Get the values assigned to those variables
-        let lhs_assignment = *witness_to_value(initial_witness, lhs_variable)?;
-        let rhs_assignment = *witness_to_value(initial_witness, rhs_variable)?;
-
-        result += coefficient * lhs_assignment * rhs_assignment;
-    }
-
-    Ok(result)
+    Ok(())
 }