+++ acvm/src/pwg/signature/ecdsa.rs
@@ -16,7 +16,7 @@ pub fn secp256k1_prehashed(
             .unwrap_or_else(|| panic!("pub_key_x should be 32 bytes long, found only {i} bytes"));
 
         let x_i = witness_to_value(initial_witness, _x_i.witness)?;
-        *pkx = *x_i.to_be_bytes().last().unwrap()
+        *pkx = *x_i.to_be_bytes().last().unwrap();
     }
 
     let mut pub_key_y = [0u8; 32];
@@ -26,7 +26,7 @@ pub fn secp256k1_prehashed(
             .unwrap_or_else(|| panic!("pub_key_y should be 32 bytes long, found only {i} bytes"));
 
         let y_i = witness_to_value(initial_witness, _y_i.witness)?;
-        *pky = *y_i.to_be_bytes().last().unwrap()
+        *pky = *y_i.to_be_bytes().last().unwrap();
     }
 
     let mut signature = [0u8; 64];
@@ -103,7 +103,7 @@ mod ecdsa_secp256k1 {
     }
 
     /// Verify an ECDSA signature, given the hashed message
-    pub fn verify_prehashed(
+    pub(super) fn verify_prehashed(
         hashed_msg: &[u8],
         public_key_x_bytes: &[u8],
         public_key_y_bytes: &[u8],